Change the sequence data type from int to short

+++ permutation.h
@@ -1,1 +1,1 @@
-int** permutation(short n, int* returnSize);
+short** permutation(short n, int* returnSize);
